colors: Split colors into files
